Added Compressed White Cobblestone, Tools, and Updated Alloy Furnace
The alloy furnace will now alloy any and all inputs but most (that are not recipes) will smelt into unusable material.
Added a new alloy called maduum which is intended to be the highest level material that actually sucks. It has a tooltip about this.
Maduum tools are indestructible, but they are absolutely terrible worse than stone tools, possibly equivalent to wood tools as they don't break and are harvest level 1.
Maduum also has an entry in the alloy furnace, any recipe that has a input of maduum nuggets, sticks, or ingots will smelt back into maduum.
Maduum will never go away.
At this point the end of 1.3.0.0 (the alloy update) is in sight as it requires the addition of armor (textures and models as well) and the textures, models, and recipes for the currently added items. 1.2.X.0 (X being the last update before the release of 1.3.0.0) will have the finshed update w/o config ability yet. The config will be the last thing to be updated.
There will be no ETA, though, as it is still a ways away.